pyproject.toml
[project]
name = "blog-converter"
version = "0.1.0"
description = "Markdown to HTML blog converter with YAML frontmatter support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "cartopy>=0.22",
    "matplotlib>=3.5",
]

[project.scripts]
blog = "converter.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["converter"]

[dependency-groups]
dev = []
